requirements.txt
# requirements.txt

# Core libraries
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3

# Network analysis libraries
networkx==2.6.2
python-igraph==0.9.7.post1

# Community detection library
python-louvain==0.15

# Progress bar library
tqdm==4.62.2

# Text processing library
Unidecode==1.3.2
